Nuked code that had old write-policy strategy in it
