+++ policy_sentry/command/write_policy.py
@@ -6,14 +6,8 @@ import json
 import logging
 import yaml
 import click
-from policy_sentry.querying.all import get_all_actions
-from policy_sentry.querying.actions import get_dependent_actions
-from policy_sentry.shared.constants import DATABASE_FILE_PATH, POLICY_LANGUAGE_VERSION
+from policy_sentry.shared.constants import DATABASE_FILE_PATH
 from policy_sentry.shared.database import connect_db
-from policy_sentry.writing.minimize import minimize_statement_actions
-from policy_sentry.writing.policy import ArnActionGroup
-from policy_sentry.writing.roles import Roles
-from policy_sentry.writing.validate import check_actions_schema, check_crud_schema
 from policy_sentry.util.file import read_yaml_file
 from policy_sentry.writing.sid_group import SidGroup
